pyproject.toml
[tool.poetry]
name = "od-compiler"
version = "0.3.2"
description = "OpenDream compiler listener"
authors = ["Crossedfall <ping@crossedfall.io>"]
maintainers = ["Crossedfall <ping@crossedfall.io>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/OpenDreamProject/od-compiler-bot"
packages = [{include = "od_compiler", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3"
docker = "^6.0"
gitpython = "^3.1"
gunicorn = "^20.1"
wget = "^3.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0"
black = "^23.3"
pre-commit = "^3.3"
codespell = "^2.2"
pytest = "^7.3"
mypy = "^1.3"
pytest-cov = "^4.0"
pytest-depends = "^1.0"
pytest-order = "^1.1"
pytest-mock = "^3.10"

[tool.poetry.scripts]
compiler = "wsgi:main"
test = "pytest:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.codespell]
count = ""
skip = ".git,.venv,*.lock,runs,OpenDream,.mypy_cache"
builtin = "clear,rare,code,en-GB_to_en-US"
ignore-words-list = "od,ro"
quiet = 3

[tool.pytest.ini_options]
addopts = "--cov=od_compiler -p no:legacypath"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
mypy_path = "od_compiler"
check_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
no_implicit_reexport = true

[[tool.mypy.overrides]]
module = "tests.*"
